--- tests/Feature/CreateApiTokenTest.php
@@ -1,46 +1,0 @@
-<?php
-
-namespace Tests\Feature;
-
-use App\Models\PersonalAccessToken;
-use App\Models\User;
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Support\Collection;
-use Laravel\Jetstream\Features;
-use Laravel\Jetstream\Http\Livewire\ApiTokenManager;
-use Livewire\Livewire;
-use Tests\TestCase;
-
-class CreateApiTokenTest extends TestCase
-{
-    use RefreshDatabase;
-
-    public function test_api_tokens_can_be_created(): void
-    {
-        if (! Features::hasApiFeatures()) {
-            $this->markTestSkipped('API support is not enabled.');
-        }
-
-        $this->actingAs($user = User::factory()->withPersonalTeam()->create());
-
-        Livewire::test(ApiTokenManager::class)
-            ->set(['createApiTokenForm' => [
-                'name' => 'Test Token',
-                'permissions' => [
-                    'read',
-                    'update',
-                ],
-            ]])
-            ->call('createApiToken');
-
-        /** @var Collection<int, PersonalAccessToken> */
-        $tokens = $user->fresh()?->tokens;
-        $this->assertCount(1, $tokens);
-
-        /** @var PersonalAccessToken */
-        $token = $tokens->first();
-        $this->assertEquals('Test Token', $token->name);
-        $this->assertTrue($token->can('read'));
-        $this->assertFalse($token->can('delete'));
-    }
-}
